+++ packages/design-system/rollup.config.js
@@ -27,7 +27,7 @@ const packageJson = require("./package.json");
 const webfont = require("postcss-webfont");
 
 export default {
-  input: "src/index.tsx",
+  input: "src/index.ts",
   output: [
     {
       file: packageJson.main,
