+++ packages/design-system/src/components/GlobalStyle/index.ts
@@ -14,8 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-export * from "./animation";
-export * from "./fonts";
-export * from "./palette";
-export * from "./spacing";
-export * from "./zindex";
+export * from "./GlobalStyle";
