+++ packages/design-system/src/index.ts
@@ -14,8 +14,22 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-export * from "./animation";
-export * from "./fonts";
-export * from "./palette";
-export * from "./spacing";
-export * from "./zindex";
+import Assets from "./assets";
+
+export { Assets };
+
+export * from "./styles";
+
+export * from "./components/Button";
+export * from "./components/Card";
+export * from "./components/ChartWrapper";
+export * from "./components/Dropdown";
+export * from "./components/ErrorPage";
+export * from "./components/GlobalStyle";
+export * from "./components/Header";
+export * from "./components/Icon";
+export * from "./components/Modal";
+export * from "./components/Need";
+export * from "./components/Search";
+export * from "./components/Tabs";
+export * from "./components/Typography";
