+++ packages/design-system/src/styles/animation.ts
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-export * from "./animation";
-export * from "./fonts";
-export * from "./palette";
-export * from "./spacing";
-export * from "./zindex";
+export const animation = {
+  defaultDurationMs: 500,
+};
